Recipes
I decided to skip on the mob_egg option this time and code that into mobSpirit by hand, as mob_egg can be used to change a spawner now, and that interferes with my implementation. If I didn't want you to have to use and empty spawner, I would just kick the whole Spirit of Mob teir away. As is, the Spirit is stealing mob_egg's render code and I will code in the spawn action if not used on a Empty Spawner.