Removed 2
I don't like it, as this is still the same mod, just version 2.0 recoded if that makes any sense.
Not SpawnerCraft2, Spawnercraft v2.
SpawnerCraft2 makes me think reworking.

Anyway, this _should_ give it back compatibility with v1 once I've added in the MobCage block. Don't quote me on that though. It's a complete re-write.