Update to using new EntityRegistry and new code used in ItemMonsterPlacer.
This still includes a call to the deprecated entityEggInfo.spawnedID. This, however, seems to be unescapeable as that is where the egg info for vanilla mobs is.

I will attempt to do a similar thing as below on the Forge mobs with the vanilla mobs by using the entityEggInfo instance but not the spawnedID such that my ItemMobSoul items use NBT rather than metadata to store all entities, instead of NBT for Forge and metadata for vanilla.

NOTE: If I'm using NBT for vanilla entities, I could then use metadata to specify the level of MobSoul. I don't think, however, that this is what I want to do: the seperate Item classes helps me keep them straight code-side.