Lazy update from metadata to NBT
The old metadata items should update to the new NBT items as soon as they look to display their name. In practice this _should_ mean they update the tick they're seen but since it's lazy it may take until it's hovered over. Either way it's a clean update process now.